Add return type annotation for Construct.sizeof
The default implementation of `Construct.sizeof` unconditionally raises a `SizeofError`, because it is meant to be overridden in subclasses. This confuses some type checkers (e.g., pyright) which incorrectly infer `Construct.sizeof` to not return (assigning the `NoReturn` return type).

Explicitly set the return type hint for `Construct.sizeof` to `int`, so that pyright correctly deals with this function.